requirements.txt
streamlit 
tensorflow
chembl_webresource_client
pandas
numpy
matplotlib
lime
google.generativeai
